Add library package an example
Closes https://github.com/guillotinaweb/guillotina_react/issues/1
Closes https://github.com/guillotinaweb/guillotina_react/issues/12

For the packaging machinery we use microbundle.
The best solution found is to provide a new dir ./package/package.json,
with the configuration for building the package.
It seems like a bit of duplication (node_modules) but just works, and
I don't want to spend more time on this.

There is also an example of integration with next.js,
Still rude, but shows the mechanic to create a ClientSide only component with
next.js. To be fully we need to provide a way of getting isomorphic-fetch
on the rest client (perhaps just adding it as dep)